+++ flask_unchained/bundles/webpack/config.py
@@ -1,14 +1,16 @@
 import os
 
-from flask_unchained import AppConfig
+from flask_unchained import BundleConfig
 
 
-class Config(AppConfig):
-    WEBPACK_MANIFEST_PATH = os.path.join(
-        AppConfig.STATIC_FOLDER, 'assets', 'manifest.json')
+class Config(BundleConfig):
+    WEBPACK_MANIFEST_PATH = (
+        None if not BundleConfig.current_app.static_folder
+        else os.path.join(BundleConfig.current_app.static_folder,
+                          'assets', 'manifest.json'))
 
 
-class ProdConfig:
+class ProdConfig(Config):
     # use relative paths by default, ie, the same host as the backend
     WEBPACK_ASSETS_HOST = ''