+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * require py-meta-utils 0.6.1 and sqlalchemy-unchained 0.5.0
 * require flask-sqlalchemy-unchained 0.4.0
 * rename `SQLAlchemy` extension class to `SQLAlchemyUnchained`
+* add a way for bundle configs to get access to the current app-under-construction
 
 ## 0.6.6 (2018/10/09)